// web/src/api/generated/models/user-detail-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface UserDetailResponse {
    'id'?: number;
    'userAccount'?: string;
    'userName'?: string;
    'userAvatar'?: string;
    'userProfile'?: string;
    'role'?: string;
    'userGender'?: number;
    'userPhone'?: string;
    'userEmail'?: string;
    'userAddress'?: string;
    'birthday'?: string;
    'level'?: number;
    'banned'?: boolean;
    'createTime'?: string;
    'updateTime'?: string;
}