// web/src/api/generated/models/update-profile-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface UpdateProfileRequest {
    'userName'?: string;
    'userAvatar'?: string;
    'userProfile'?: string;
    'userGender'?: number;
    'userPhone'?: string;
    'phoneSmsCode'?: string;
    'userEmail'?: string;
    'userAddress'?: string;
    'birthday'?: string;
}